Corrección de Gastos, Productos y Salario (Integer a Double), corrección de datos de Informes, corrección de muestra de datos en pantalla

+++ Escritorio/src/Compartir/Usuarios.java
@@ -14,7 +14,7 @@ public class Usuarios implements Serializable {
     private String usuario;
     private String nombre;
     private String contraseña;
-    private int salario;
+    private double salario;
     private String imagen;
     private boolean existe;
 /**
@@ -32,7 +32,7 @@ public class Usuarios implements Serializable {
  * @param imagen imagen codificada del usuario
  * @param existe boolean para indicar si existe un usuario
  */
-    public Usuarios(int id, String usuario, String nombre, String contraseña, int salario, String imagen, boolean existe) {
+    public Usuarios(int id, String usuario, String nombre, String contraseña, double salario, String imagen, boolean existe) {
         this.id = id;
         this.usuario = usuario;
         this.nombre = nombre;
@@ -80,13 +80,14 @@ public class Usuarios implements Serializable {
         this.contraseña = contraseña;
     }
 
-    public int getSalario() {
+    public double getSalario() {
         return salario;
     }
 
-    public void setSalario(int salario) {
+    public void setSalario(double salario) {
         this.salario = salario;
     }
+
 
     public boolean isExiste() {
         return existe;
